create kafka_app and consolidate kafka-related pieces there
