+++ django/src/kafka_app/consumer.py
