+++ django/src/kafka_app/tests/mock.py
